Created ErrorMessage class and error messages in DAOs are now standardized

+++ src/main/java/piiksuma/api/dao/MultimediaDao.java
@@ -2,6 +2,7 @@ package piiksuma.api.dao;
 
 import piiksuma.Multimedia;
 import piiksuma.Post;
+import piiksuma.api.ErrorMessage;
 import piiksuma.database.DeleteMapper;
 import piiksuma.database.InsertionMapper;
 import piiksuma.database.QueryMapper;
@@ -19,7 +20,7 @@ public class MultimediaDao extends AbstractDao {
 
     public void addMultimedia(Multimedia multimedia) throws PiikDatabaseException {
         if (multimedia == null || !multimedia.checkPrimaryKey()) {
-            throw new PiikDatabaseException("(multimedia) Primary key constraints failed");
+            throw new PiikDatabaseException(ErrorMessage.getPkConstraintMessage("multimedia"));
         }
 
         new InsertionMapper<Multimedia>(super.getConnection()).add(multimedia).defineClass(Multimedia.class).insert();
@@ -33,6 +34,10 @@ public class MultimediaDao extends AbstractDao {
      * @throws PiikDatabaseException
      */
     public Multimedia existsMultimedia(Multimedia multimedia) throws PiikDatabaseException {
+        if (multimedia == null || !multimedia.checkPrimaryKey()) {
+            throw new PiikDatabaseException(ErrorMessage.getPkConstraintMessage("multimedia"));
+        }
+
         return new QueryMapper<Multimedia>(super.getConnection()).createQuery("SELECT * FROM multimedia" +
                 "WHERE hash=?").defineClass(Multimedia.class).defineParameters(multimedia.getHash()).findFirst();
     }
@@ -43,14 +48,24 @@ public class MultimediaDao extends AbstractDao {
      * @param multimedia Multimedia about which we want to know on how many posts is included
      * @return Number of post which contains the multimedia
      */
-    public Long numPostMultimedia(Multimedia multimedia) {
+    public Long numPostMultimedia(Multimedia multimedia) throws PiikDatabaseException {
+        if (multimedia == null || !multimedia.checkPrimaryKey()) {
+            throw new PiikDatabaseException(ErrorMessage.getPkConstraintMessage("multimedia"));
+        }
+
         List<Map<String, Object>> count = new QueryMapper<Object>(super.getConnection()).createQuery("SELECT COUNT(*) "+
                 "as numPostMultimedia FROM post WHERE multimedia = ?").defineClass(Object.class)
                 .defineParameters(multimedia.getHash()).mapList();
+
         return (Long) count.get(0).get("numPostMultimedia");
     }
 
-    public List<Post> postWithMultimedia(Multimedia multimedia) {
+    public List<Post> postWithMultimedia(Multimedia multimedia) throws PiikDatabaseException {
+
+        if (multimedia == null || !multimedia.checkPrimaryKey()) {
+            throw new PiikDatabaseException(ErrorMessage.getPkConstraintMessage("multimedia"));
+        }
+
         return null;
     }
 
@@ -61,6 +76,10 @@ public class MultimediaDao extends AbstractDao {
      * @throws PiikDatabaseException
      */
     public void removeMultimedia(Multimedia multimedia) throws PiikDatabaseException {
+
+        if (multimedia == null || !multimedia.checkPrimaryKey()) {
+            throw new PiikDatabaseException(ErrorMessage.getPkConstraintMessage("multimedia"));
+        }
 
         new DeleteMapper<Multimedia>(super.getConnection()).add(multimedia).defineClass(Multimedia.class).delete();
     }
